No rotations, boss added
The same game with some changes without any rotation to make things
easier.

After a certain score is reached a boss will appear.

There are now nice images to represent the asteroids.

To do:

The way of killing the boss is still incomplete, the score behaves in an
odd way.

Asteroids should probably split into two